Client: now handles session timeout
1. Configured an AngularJS http interceptor for response errors. When a
response error happens due to a 401, the user is sent to the login page.

2. Removed resolving of the current user in all ui router states since
it is not required. Any requests going to the server will have the
X-XSRF-TOKEN header set in the HTTP request, if it is invalid the server
will respond with a 401. The interceptor now takes the user to the login
page for a 401 response.