Modules now include their dependencies
A new module called "base" was created and services from the common module were moved to this new module.

The module hierarchy is now:

1. base
2. domain
3. common
4. centres, studies, users (these don't depend on each other )
5. admin, collection, shipmentSpecimens (these don't depend on each other )

Where each of the modules depend on the lower numbered ones in the list.

Test suites now depend on the module the component / directvie / service is defined in. This makes test run
faster since the whole application is not loaded for every test.